+++ src/components/heading/index.ts
@@ -1,0 +1,1 @@
+export { default as VfHeading } from "./VfHeading.vue";