+++ src/components/prose/index.ts
@@ -1,0 +1,1 @@
+export { default as VfProse } from "./VfProse.vue";