+++ src/components/text/index.ts
@@ -1,0 +1,1 @@
+export { default as VfText } from "./VfText.vue";